Move clear cookies key out of meta and move tests into other integration tests
