+++ tavern/_plugins/rest/request.py
@@ -383,10 +383,8 @@ class RestRequest(BaseRequest):
                 spec. Only valid keyword args to requests can be passed
         """
 
-        if "meta" in rspec:
-            meta = rspec.pop("meta")
-            if meta and "clear_session_cookies" in meta:
-                session.cookies.clear_session_cookies()
+        if rspec.pop("clear_session_cookies", False):
+            session.cookies.clear_session_cookies()
 
         expected = {
             "method",
